Add G1 points in monomial form to the trusted setup
This commit will speed up loading the trusted setup by about ~1.5 seconds, the
time it takes to convert the G1 points in Lagrange form to monomial form. These
values were copied from the trusted setup in the consensus specs.

We added the new points to the bottom of the trusted setup so that it is
backwards compatiable with previous versions of c-kzg-4844. So clients could
start using this trusted setup whenever they want.